change callback to uint32

+++ tasmota/tasmota_xdrv_driver/xdrv_02_9_mqtt.ino
@@ -1984,7 +1984,7 @@ void MqttSaveSettings(void) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv02(uint8_t function)
+bool Xdrv02(uint32_t function)
 {
   bool result = false;
