+++ tasmota/tasmota_xdrv_driver/xdrv_03_energy.ino
@@ -1337,7 +1337,7 @@ void EnergyShow(bool json) {
  * Interface
 \*********************************************************************************************/
 
-bool Xdrv03(uint8_t function)
+bool Xdrv03(uint32_t function)
 {
   bool result = false;
 
@@ -1373,7 +1373,7 @@ bool Xdrv03(uint8_t function)
   return result;
 }
 
-bool Xsns03(uint8_t function)
+bool Xsns03(uint32_t function)
 {
   bool result = false;
